move stories into separat directory
