+++ wls-gui-wahllokalsystem/stories/components/common/buttons/BaseButtonSave.stories.ts
@@ -1,15 +1,21 @@
 import type { Meta, StoryObj } from "@storybook/vue3";
 
-import BaseButtonRefresh from "./BaseButtonRefresh.vue";
+import BaseButtonSave from "../../../../src/components/common/buttons/BaseButtonSave.vue";
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories
 const meta = {
-  component: BaseButtonRefresh,
+  component: BaseButtonSave,
   args: {},
-} satisfies Meta<typeof BaseButtonRefresh>;
+} satisfies Meta<typeof BaseButtonSave>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 export const Default: Story = {
   args: {},
+};
+
+export const Disabled: Story = {
+  args: {
+    disabled: true,
+  },
 };
